AML-3881: move next test across. Also bring over extension helper and fixup namespace in existing C&P extensions.
