+++ src/SFA.DAS.EmployerAccounts.Api.IntegrationTests/GivenEmployerAccountsApi/LegalEntitiesControllerTests/WhenIGetASingleLegalEntityWithKnownIds.cs
@@ -1,9 +1,8 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using NUnit.Framework;
-using SFA.DAS.EAS.Account.API.IntegrationTests.TestUtils.DataAccess.DataHelpers;
+using SFA.DAS.EmployerAccounts.Api.IntegrationTests.Helpers;
 using SFA.DAS.EmployerAccounts.Api.IntegrationTests.ModelBuilders;
-using SFA.DAS.EmployerAccounts.Api.IntegrationTests.TestUtils.DataAccess;
 using SFA.DAS.EmployerAccounts.Api.IntegrationTests.TestUtils.DataAccess.Dtos;
 using SFA.DAS.EmployerAccounts.Api.Types;
